+++ R/RcppExports.R
@@ -13,6 +13,10 @@ xlsx_validation_ <- function(path, sheet_paths, sheet_names) {
     .Call('_tidyxl_xlsx_validation_', PACKAGE = 'tidyxl', path, sheet_paths, sheet_names)
 }
 
+xlsx_names_ <- function(path) {
+    .Call('_tidyxl_xlsx_names_', PACKAGE = 'tidyxl', path)
+}
+
 xlex_ <- function(x) {
     .Call('_tidyxl_xlex_', PACKAGE = 'tidyxl', x)
 }